restructure UI classes
